Revert "see if PyPy error moves"
This reverts commit 8654141943785cdac996b4a913c6e29ceccb4dcd.
